--- a/benchmark/ICSharpCode.SharpZipLib.Benchmark/Program.cs
+++ b/benchmark/ICSharpCode.SharpZipLib.Benchmark/Program.cs
@@ -1,9 +1,8 @@
-﻿using System;
-using BenchmarkDotNet;
-using BenchmarkDotNet.Configs;
+﻿using BenchmarkDotNet.Configs;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Running;
 using BenchmarkDotNet.Toolchains.CsProj;
+using ICSharpCode.SharpZipLib.Benchmark.Tar;
 
 namespace ICSharpCode.SharpZipLib.Benchmark
 {
@@ -22,6 +21,12 @@ namespace ICSharpCode.SharpZipLib.Benchmark
 		static void Main(string[] args)
 		{
 			BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+			
+			// var output = new TarOutputStream();
+			// for (int i = 0; i < 1_000_000; i++)
+			// {
+			// 	output.WriteTarOutputStream();
+			// }
 		}
 	}
 }