--- a/src/ICSharpCode.SharpZipLib/Tar/TarBuffer.cs
+++ b/src/ICSharpCode.SharpZipLib/Tar/TarBuffer.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Buffers;
 using System.IO;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace ICSharpCode.SharpZipLib.Tar
 {
@@ -73,10 +75,7 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// This is equal to the <see cref="BlockFactor"/> multiplied by the <see cref="BlockSize"/></value>
 		public int RecordSize
 		{
-			get
-			{
-				return recordSize;
-			}
+			get { return recordSize; }
 		}
 
 		/// <summary>
@@ -96,10 +95,7 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// <value>This is the number of blocks in each record.</value>
 		public int BlockFactor
 		{
-			get
-			{
-				return blockFactor;
-			}
+			get { return blockFactor; }
 		}
 
 		/// <summary>
@@ -292,6 +288,19 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </summary>
 		public void SkipBlock()
 		{
+			SkipBlockAsync(CancellationToken.None, false).GetAwaiter().GetResult();
+		}
+
+		/// <summary>
+		/// Skip over a block on the input stream.
+		/// </summary>
+		public Task SkipBlockAsync(CancellationToken ct)
+		{
+			return SkipBlockAsync(ct, true).AsTask();
+		}
+
+		private async ValueTask SkipBlockAsync(CancellationToken ct, bool isAsync)
+		{
 			if (inputStream == null)
 			{
 				throw new TarException("no input stream defined");
@@ -299,7 +308,7 @@ namespace ICSharpCode.SharpZipLib.Tar
 
 			if (currentBlockIndex >= BlockFactor)
 			{
-				if (!ReadRecord())
+				if (!await ReadRecordAsync(ct, isAsync))
 				{
 					throw new TarException("Failed to read a record");
 				}
@@ -323,7 +332,7 @@ namespace ICSharpCode.SharpZipLib.Tar
 
 			if (currentBlockIndex >= BlockFactor)
 			{
-				if (!ReadRecord())
+				if (!ReadRecordAsync(CancellationToken.None, false).GetAwaiter().GetResult())
 				{
 					throw new TarException("Failed to read a record");
 				}
@@ -335,14 +344,14 @@ namespace ICSharpCode.SharpZipLib.Tar
 			currentBlockIndex++;
 			return result;
 		}
-		
-		internal void ReadBlockInt(Span<byte> buffer)
+
+		internal async ValueTask ReadBlockIntAsync(byte[] buffer, CancellationToken ct, bool isAsync)
 		{
 			if (buffer.Length != BlockSize)
 			{
 				throw new ArgumentException("BUG: buffer must have length BlockSize");
 			}
-			
+
 			if (inputStream == null)
 			{
 				throw new TarException("TarBuffer.ReadBlock - no input stream defined");
@@ -350,13 +359,13 @@ namespace ICSharpCode.SharpZipLib.Tar
 
 			if (currentBlockIndex >= BlockFactor)
 			{
-				if (!ReadRecord())
+				if (!await ReadRecordAsync(ct, isAsync))
 				{
 					throw new TarException("Failed to read a record");
 				}
 			}
-			
-			recordBuffer.AsSpan().Slice(currentBlockIndex* BlockSize, BlockSize).CopyTo(buffer);
+
+			recordBuffer.AsSpan().Slice(currentBlockIndex * BlockSize, BlockSize).CopyTo(buffer);
 			currentBlockIndex++;
 		}
 
@@ -366,7 +375,7 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// <returns>
 		/// false if End-Of-File, else true.
 		/// </returns>
-		private bool ReadRecord()
+		private async ValueTask<bool> ReadRecordAsync(CancellationToken ct, bool isAsync)
 		{
 			if (inputStream == null)
 			{
@@ -380,7 +389,9 @@ namespace ICSharpCode.SharpZipLib.Tar
 
 			while (bytesNeeded > 0)
 			{
-				long numBytes = inputStream.Read(recordBuffer, offset, bytesNeeded);
+				long numBytes = isAsync
+					? await inputStream.ReadAsync(recordBuffer, offset, bytesNeeded, ct)
+					: inputStream.Read(recordBuffer, offset, bytesNeeded);
 
 				//
 				// NOTE
@@ -469,32 +480,38 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// <param name="block">
 		/// The data to write to the archive.
 		/// </param>
+		/// <param name="ct"></param>
+		public ValueTask WriteBlockAsync(byte[] block, CancellationToken ct)
+		{
+			return WriteBlockAsync(block, 0, ct);
+		}
+		
+		/// <summary>
+		/// Write a block of data to the archive.
+		/// </summary>
+		/// <param name="block">
+		/// The data to write to the archive.
+		/// </param>
 		public void WriteBlock(byte[] block)
 		{
-			if (block == null)
-			{
-				throw new ArgumentNullException(nameof(block));
-			}
+			WriteBlock(block, 0);
+		}
 
-			if (outputStream == null)
-			{
-				throw new TarException("TarBuffer.WriteBlock - no output stream defined");
-			}
-
-			if (block.Length != BlockSize)
-			{
-				string errorText = string.Format("TarBuffer.WriteBlock - block to write has length '{0}' which is not the block size of '{1}'",
-					block.Length, BlockSize);
-				throw new TarException(errorText);
-			}
-
-			if (currentBlockIndex >= BlockFactor)
-			{
-				WriteRecord();
-			}
-
-			Array.Copy(block, 0, recordBuffer, (currentBlockIndex * BlockSize), BlockSize);
-			currentBlockIndex++;
+		/// <summary>
+		/// Write an archive record to the archive, where the record may be
+		/// inside of a larger array buffer. The buffer must be "offset plus
+		/// record size" long.
+		/// </summary>
+		/// <param name="buffer">
+		/// The buffer containing the record data to write.
+		/// </param>
+		/// <param name="offset">
+		/// The offset of the record data within buffer.
+		/// </param>
+		/// <param name="ct"></param>
+		public ValueTask WriteBlockAsync(byte[] buffer, int offset, CancellationToken ct)
+		{
+			return WriteBlockAsync(buffer, offset, ct, true);
 		}
 
 		/// <summary>
@@ -509,6 +526,11 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// The offset of the record data within buffer.
 		/// </param>
 		public void WriteBlock(byte[] buffer, int offset)
+		{
+			WriteBlockAsync(buffer, offset, CancellationToken.None, false).GetAwaiter().GetResult();
+		}
+
+		internal async ValueTask WriteBlockAsync(byte[] buffer, int offset, CancellationToken ct, bool isAsync)
 		{
 			if (buffer == null)
 			{
@@ -527,14 +549,15 @@ namespace ICSharpCode.SharpZipLib.Tar
 
 			if ((offset + BlockSize) > buffer.Length)
 			{
-				string errorText = string.Format("TarBuffer.WriteBlock - record has length '{0}' with offset '{1}' which is less than the record size of '{2}'",
+				string errorText = string.Format(
+					"TarBuffer.WriteBlock - record has length '{0}' with offset '{1}' which is less than the record size of '{2}'",
 					buffer.Length, offset, recordSize);
 				throw new TarException(errorText);
 			}
 
 			if (currentBlockIndex >= BlockFactor)
 			{
-				WriteRecord();
+				await WriteRecordAsync(CancellationToken.None, isAsync);
 			}
 
 			Array.Copy(buffer, offset, recordBuffer, (currentBlockIndex * BlockSize), BlockSize);
@@ -545,15 +568,23 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// <summary>
 		/// Write a TarBuffer record to the archive.
 		/// </summary>
-		private void WriteRecord()
+		private async ValueTask WriteRecordAsync(CancellationToken ct, bool isAsync)
 		{
 			if (outputStream == null)
 			{
 				throw new TarException("TarBuffer.WriteRecord no output stream defined");
 			}
 
-			outputStream.Write(recordBuffer, 0, RecordSize);
-			outputStream.Flush();
+			if (isAsync)
+			{
+				await outputStream.WriteAsync(recordBuffer, 0, RecordSize, ct);
+				await outputStream.FlushAsync(ct);
+			}
+			else
+			{
+				outputStream.Write(recordBuffer, 0, RecordSize);
+				outputStream.Flush();
+			}
 
 			currentBlockIndex = 0;
 			currentRecordIndex++;
@@ -564,7 +595,7 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </summary>
 		/// <remarks>Any trailing bytes are set to zero which is by definition correct behaviour
 		/// for the end of a tar stream.</remarks>
-		private void WriteFinalRecord()
+		private async ValueTask WriteFinalRecordAsync(CancellationToken ct, bool isAsync)
 		{
 			if (outputStream == null)
 			{
@@ -575,36 +606,76 @@ namespace ICSharpCode.SharpZipLib.Tar
 			{
 				int dataBytes = currentBlockIndex * BlockSize;
 				Array.Clear(recordBuffer, dataBytes, RecordSize - dataBytes);
-				WriteRecord();
+				await WriteRecordAsync(ct, isAsync);
 			}
 
-			outputStream.Flush();
+			if (isAsync)
+			{
+				await outputStream.FlushAsync(ct);
+			}
+			else
+			{
+				outputStream.Flush();
+			}
 		}
 
 		/// <summary>
 		/// Close the TarBuffer. If this is an output buffer, also flush the
 		/// current block before closing.
 		/// </summary>
-		public void Close()
+		public void Close() => CloseAsync(CancellationToken.None, false).GetAwaiter().GetResult();
+		
+		/// <summary>
+		/// Close the TarBuffer. If this is an output buffer, also flush the
+		/// current block before closing.
+		/// </summary>
+		public Task CloseAsync(CancellationToken ct) => CloseAsync(ct, true).AsTask();
+
+		private async ValueTask CloseAsync(CancellationToken ct, bool isAsync)
 		{
 			if (outputStream != null)
 			{
-				WriteFinalRecord();
+				await WriteFinalRecordAsync(ct, isAsync);
 
 				if (IsStreamOwner)
 				{
-					outputStream.Dispose();
+					if (isAsync)
+					{
+#if NETSTANDARD2_1_OR_GREATER
+						await outputStream.DisposeAsync();
+#else
+						outputStream.Dispose();
+#endif
+					}
+					else
+					{
+						outputStream.Dispose();
+					}
 				}
+
 				outputStream = null;
 			}
 			else if (inputStream != null)
 			{
 				if (IsStreamOwner)
 				{
-					inputStream.Dispose();
+					if (isAsync)
+					{
+#if NETSTANDARD2_1_OR_GREATER
+						await inputStream.DisposeAsync();
+#else
+						inputStream.Dispose();
+#endif
+					}
+					else
+					{
+						inputStream.Dispose();
+					}
 				}
+
 				inputStream = null;
 			}
+
 			ArrayPool<byte>.Shared.Return(recordBuffer);
 		}