--- a/src/ICSharpCode.SharpZipLib/Tar/TarEntry.cs
+++ b/src/ICSharpCode.SharpZipLib/Tar/TarEntry.cs
@@ -114,7 +114,8 @@ namespace ICSharpCode.SharpZipLib.Tar
 		public static TarEntry CreateTarEntry(string name)
 		{
 			var entry = new TarEntry();
-			TarEntry.NameTarHeader(entry.header, name);
+
+			entry.NameTarHeader(name);
 			return entry;
 		}
 
@@ -188,10 +189,7 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </returns>
 		public TarHeader TarHeader
 		{
-			get
-			{
-				return header;
-			}
+			get { return header; }
 		}
 
 		/// <summary>
@@ -199,14 +197,8 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </summary>
 		public string Name
 		{
-			get
-			{
-				return header.Name;
-			}
-			set
-			{
-				header.Name = value;
-			}
+			get { return header.Name; }
+			set { header.Name = value; }
 		}
 
 		/// <summary>
@@ -214,14 +206,8 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </summary>
 		public int UserId
 		{
-			get
-			{
-				return header.UserId;
-			}
-			set
-			{
-				header.UserId = value;
-			}
+			get { return header.UserId; }
+			set { header.UserId = value; }
 		}
 
 		/// <summary>
@@ -229,14 +215,8 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </summary>
 		public int GroupId
 		{
-			get
-			{
-				return header.GroupId;
-			}
-			set
-			{
-				header.GroupId = value;
-			}
+			get { return header.GroupId; }
+			set { header.GroupId = value; }
 		}
 
 		/// <summary>
@@ -244,14 +224,8 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </summary>
 		public string UserName
 		{
-			get
-			{
-				return header.UserName;
-			}
-			set
-			{
-				header.UserName = value;
-			}
+			get { return header.UserName; }
+			set { header.UserName = value; }
 		}
 
 		/// <summary>
@@ -259,14 +233,8 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </summary>
 		public string GroupName
 		{
-			get
-			{
-				return header.GroupName;
-			}
-			set
-			{
-				header.GroupName = value;
-			}
+			get { return header.GroupName; }
+			set { header.GroupName = value; }
 		}
 
 		/// <summary>
@@ -304,14 +272,8 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </summary>
 		public DateTime ModTime
 		{
-			get
-			{
-				return header.ModTime;
-			}
-			set
-			{
-				header.ModTime = value;
-			}
+			get { return header.ModTime; }
+			set { header.ModTime = value; }
 		}
 
 		/// <summary>
@@ -322,10 +284,7 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </returns>
 		public string File
 		{
-			get
-			{
-				return file;
-			}
+			get { return file; }
 		}
 
 		/// <summary>
@@ -333,14 +292,8 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// </summary>
 		public long Size
 		{
-			get
-			{
-				return header.Size;
-			}
-			set
-			{
-				header.Size = value;
-			}
+			get { return header.Size; }
+			set { header.Size = value; }
 		}
 
 		/// <summary>
@@ -450,7 +403,8 @@ namespace ICSharpCode.SharpZipLib.Tar
 				header.Size = new FileInfo(file.Replace('/', Path.DirectorySeparatorChar)).Length;
 			}
 
-			header.ModTime = System.IO.File.GetLastWriteTime(file.Replace('/', Path.DirectorySeparatorChar)).ToUniversalTime();
+			header.ModTime = System.IO.File.GetLastWriteTime(file.Replace('/', Path.DirectorySeparatorChar))
+				.ToUniversalTime();
 			header.DevMajor = 0;
 			header.DevMinor = 0;
 		}
@@ -549,13 +503,8 @@ namespace ICSharpCode.SharpZipLib.Tar
 		/// <param name="name">
 		/// The tar entry name.
 		/// </param>
-		static public void NameTarHeader(TarHeader header, string name)
+		public void NameTarHeader(string name)
 		{
-			if (header == null)
-			{
-				throw new ArgumentNullException(nameof(header));
-			}
-
 			if (name == null)
 			{
 				throw new ArgumentNullException(nameof(name));