… src/ICSharpCode.SharpZipLib/Tar/TarHeader.cs
… test/ICSharpCode.SharpZipLib.Tests/Core/StringBuilderPoolTests.cs
… test/ICSharpCode.SharpZipLib.Tests/Tar/TarBufferTests.cs
… test/ICSharpCode.SharpZipLib.Tests/Tar/TarInputStreamTests.cs
… test/ICSharpCode.SharpZipLib.Tests/Tar/TarTests.cs