<!---
Please remember that unless we have a Joint Copyright Agreement on file or the following statement is in your pull request, we cannot accept it.
-->
_I certify that I own, and have sufficient rights to contribute, all source code and related material intended to be compiled or integrated with the source code for the SharpZipLib open source product (the "Contribution"). My Contribution is licensed under the MIT License._

Relates to #223 

I have taken a first pass at making TAR support async streams which will help with streaming to and from network streams. I have also tried to dramatically reduce memory allocations associated with reading from and writing to TAR's.

I haven't added new tests yet or really checked what the coverage is like but if you are happy with how it looks as it, I can do that.

Below are some benchmarks run before and after my changes:  
Before:  
|                  Method |        Job |     Toolchain |       Mean |    Error |   StdDev |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------------ |----------- |-------------- |-----------:|---------:|---------:|--------:|-------:|------:|----------:|
|      ReadTarInputStream | Job-IFGTKW | .NET Core 3.1 |   122.3 μs |  1.71 μs |  1.43 μs | 33.2031 | 1.2207 |     - | 1112144 B |
| ReadTarInputStreamAsync | Job-IFGTKW | .NET Core 3.1 | 1,081.9 μs | 21.64 μs | 58.50 μs | 37.1094 | 0.9766 |     - | 1227248 B |
|      WriteTarOutputStream | Job-IFGTKW | .NET Core 3.1 |    50.54 μs |  0.339 μs |  0.283 μs | 0.3052 |     - |     - |   12008 B |
| WriteTarOutputStreamAsync | Job-IFGTKW | .NET Core 3.1 | 1,135.09 μs | 22.460 μs | 51.154 μs | 3.9063 |     - |     - |  126985 B |

After:  
|                  Method |        Job |     Toolchain |     Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------ |----------- |-------------- |---------:|--------:|--------:|------:|------:|------:|----------:|
|      ReadTarInputStream | Job-KKAMPY | .NET Core 3.1 | 221.3 μs | 0.76 μs | 0.64 μs |     - |     - |     - |     952 B |
| ReadTarInputStreamAsync | Job-KKAMPY | .NET Core 3.1 | 227.1 μs | 4.05 μs | 3.79 μs |     - |     - |     - |    1096 B |
|      WriteTarOutputStream | Job-KKAMPY | .NET Core 3.1 | 149.7 μs | 2.71 μs | 2.54 μs |     - |     - |     - |     449 B |
| WriteTarOutputStreamAsync | Job-KKAMPY | .NET Core 3.1 | 146.7 μs | 0.67 μs | 0.56 μs |     - |     - |     - |     448 B |