@piksel 2022-04-11
Wow, for some reason I thought `Stream.WriteAsync` was not available in .NET 4.5... 
This should make the `async` versions of `Zip[In/Out]putStream` much clearer.

Anyway, this mostly looks great, but it does lower the test coverage a bit.