@nathan-c 2022-04-11
Thanks for the feedback! I will create some more tests to cover the changes